Clean up JS test utils structure
